+++ api/src/main/java/org/allaymc/api/entity/component/EntityFireworksRocketBaseComponent.java
@@ -1,7 +1,7 @@
 package org.allaymc.api.entity.component;
 
 import org.allaymc.api.entity.interfaces.EntityPlayer;
-import org.allaymc.api.world.FireworkExplosion;
+import org.allaymc.api.world.explosion.FireworkExplosion;
 import org.jetbrains.annotations.UnmodifiableView;
 
 import java.util.Set;